refactor(providers): extract shared endpoint schemas to decorator-free modules
The flat-file collapse merged each endpoint's pydantic models into its
endpoint file, next to the @register decorator. An endpoint file that
imports a shared schema from a sibling (exa contents/find_similar ->
search's Contents; firecrawl crawl -> scrape's OutputFormat) therefore
fired the sibling's registration on lazy attribute access, defeating
import laziness.

Move the shared schemas into decorator-free _schemas.py modules so a
lazy attr access loads only its own endpoint module plus the shared
schema, never a sibling Endpoint class. The full provider test/import
suite is unchanged (525 passed) and per-attribute laziness now holds.

Also complete each provider's _LAZY_MAP and module __all__ with the
primary request classes that were omitted (ExaSearchRequest,
PerplexityChatRequest, TavilySearchRequest, TavilyExtractRequest) so
every endpoint and its request class is re-exported as documented, and
fix a stale "subdirectories are capabilities" line in the provider docs.
