+++ lionagi/providers/exa/find_similar.py
@@ -7,7 +7,7 @@ from __future__ import annotations
 
 from pydantic import BaseModel, ConfigDict, Field
 
-from lionagi.providers.exa.search import Contents
+from lionagi.providers.exa._schemas import Contents
 from lionagi.service.connections.endpoint import Endpoint
 from lionagi.service.connections.endpoint_config import EndpointConfig
