+++ lionagi/providers/perplexity/__init__.py
@@ -5,6 +5,7 @@ from lionagi.ln._lazy_init import lazy_import
 
 _LAZY_MAP: dict[str, tuple[str, str | None]] = {
     "PerplexityChatEndpoint": ("chat", None),
+    "PerplexityChatRequest": ("chat", None),
 }
